+++ .eslintcache
@@ -1,1 +1,1 @@
-[{"D:\\www\\kinopoisk\\src\\reportWebVitals.js":"1","D:\\www\\kinopoisk\\src\\App.js":"2","D:\\www\\kinopoisk\\src\\Movies\\MoviesList.jsx":"3","D:\\www\\kinopoisk\\src\\Movies\\MovieItem.jsx":"4","D:\\www\\kinopoisk\\src\\Filters\\Filters.jsx":"5","D:\\www\\kinopoisk\\src\\config\\api.js":"6","D:\\www\\kinopoisk\\src\\Filters\\SortBy.jsx":"7","D:\\www\\kinopoisk\\src\\index.js":"8","D:\\www\\kinopoisk\\src\\Filters\\Year.jsx":"9","D:\\www\\kinopoisk\\src\\Filters\\Genres.jsx":"10"},{"size":362,"mtime":1609800407800,"results":"11","hashOfConfig":"12"},{"size":1553,"mtime":1611622737794,"results":"13","hashOfConfig":"12"},{"size":1468,"mtime":1611622947469,"results":"14","hashOfConfig":"12"},{"size":667,"mtime":1611537454089,"results":"15","hashOfConfig":"12"},{"size":1193,"mtime":1611619261295,"results":"16","hashOfConfig":"12"},{"size":356,"mtime":1611014942433,"results":"17","hashOfConfig":"12"},{"size":1160,"mtime":1611537008897,"results":"18","hashOfConfig":"12"},{"size":500,"mtime":1609800407800,"results":"19","hashOfConfig":"12"},{"size":711,"mtime":1611565767578,"results":"20","hashOfConfig":"12"},{"size":1132,"mtime":1611700644468,"results":"21","hashOfConfig":"12"},{"filePath":"22","messages":"23","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"7ebhbf",{"filePath":"24","messages":"25","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"26","messages":"27","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"28","messages":"29","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"30","messages":"31","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"32","messages":"33","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"34","messages":"35","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"36","messages":"37","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"38","messages":"39","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"40","messages":"41","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"D:\\www\\kinopoisk\\src\\reportWebVitals.js",[],"D:\\www\\kinopoisk\\src\\App.js",[],"D:\\www\\kinopoisk\\src\\Movies\\MoviesList.jsx",[],"D:\\www\\kinopoisk\\src\\Movies\\MovieItem.jsx",[],"D:\\www\\kinopoisk\\src\\Filters\\Filters.jsx",[],"D:\\www\\kinopoisk\\src\\config\\api.js",[],"D:\\www\\kinopoisk\\src\\Filters\\SortBy.jsx",[],"D:\\www\\kinopoisk\\src\\index.js",[],"D:\\www\\kinopoisk\\src\\Filters\\Year.jsx",[],"D:\\www\\kinopoisk\\src\\Filters\\Genres.jsx",[]]
+[{"D:\\www\\kinopoisk\\src\\reportWebVitals.js":"1","D:\\www\\kinopoisk\\src\\App.js":"2","D:\\www\\kinopoisk\\src\\Movies\\MoviesList.jsx":"3","D:\\www\\kinopoisk\\src\\Movies\\MovieItem.jsx":"4","D:\\www\\kinopoisk\\src\\Filters\\Filters.jsx":"5","D:\\www\\kinopoisk\\src\\config\\api.js":"6","D:\\www\\kinopoisk\\src\\Filters\\SortBy.jsx":"7","D:\\www\\kinopoisk\\src\\index.js":"8","D:\\www\\kinopoisk\\src\\Filters\\Year.jsx":"9","D:\\www\\kinopoisk\\src\\Filters\\Genres.jsx":"10","D:\\www\\kinopoisk\\src\\pages\\Movies.js":"11","D:\\www\\kinopoisk\\src\\pages\\Home.js":"12","D:\\www\\kinopoisk\\src\\pages\\Movie.js":"13"},{"size":362,"mtime":1609800407800,"results":"14","hashOfConfig":"15"},{"size":1461,"mtime":1611875811927,"results":"16","hashOfConfig":"15"},{"size":1468,"mtime":1611622947469,"results":"17","hashOfConfig":"15"},{"size":667,"mtime":1611537454089,"results":"18","hashOfConfig":"15"},{"size":1193,"mtime":1611619261295,"results":"19","hashOfConfig":"15"},{"size":356,"mtime":1611014942433,"results":"20","hashOfConfig":"15"},{"size":1160,"mtime":1611537008897,"results":"21","hashOfConfig":"15"},{"size":500,"mtime":1609800407800,"results":"22","hashOfConfig":"15"},{"size":711,"mtime":1611565767578,"results":"23","hashOfConfig":"15"},{"size":1132,"mtime":1611700644468,"results":"24","hashOfConfig":"15"},{"size":1202,"mtime":1611874070541,"results":"25","hashOfConfig":"15"},{"size":181,"mtime":1611874033043,"results":"26","hashOfConfig":"15"},{"size":190,"mtime":1611875080248,"results":"27","hashOfConfig":"15"},{"filePath":"28","messages":"29","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"7ebhbf",{"filePath":"30","messages":"31","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"32","messages":"33","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"34","messages":"35","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"36","messages":"37","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"38","messages":"39","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"40","messages":"41","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"42","messages":"43","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"44","messages":"45","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"46","messages":"47","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"48","messages":"49","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"50","messages":"51","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"52","messages":"53","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"D:\\www\\kinopoisk\\src\\reportWebVitals.js",[],"D:\\www\\kinopoisk\\src\\App.js",[],"D:\\www\\kinopoisk\\src\\Movies\\MoviesList.jsx",[],"D:\\www\\kinopoisk\\src\\Movies\\MovieItem.jsx",[],"D:\\www\\kinopoisk\\src\\Filters\\Filters.jsx",[],"D:\\www\\kinopoisk\\src\\config\\api.js",[],"D:\\www\\kinopoisk\\src\\Filters\\SortBy.jsx",[],"D:\\www\\kinopoisk\\src\\index.js",[],"D:\\www\\kinopoisk\\src\\Filters\\Year.jsx",[],"D:\\www\\kinopoisk\\src\\Filters\\Genres.jsx",[],"D:\\www\\kinopoisk\\src\\pages\\Movies.js",[],"D:\\www\\kinopoisk\\src\\pages\\Home.js",[],"D:\\www\\kinopoisk\\src\\pages\\Movie.js",[]]
