+++ src/index.css
@@ -17,3 +17,11 @@ code {
   width: 100%;
   height: 400px;
 }
+
+.logo {
+  color: #fff;
+  font-size: 24px;
+  font-weight: bold;
+  margin-top: -4px;
+  margin-right: 20px;
+}
