+++ src/pages/Home.js
@@ -1,0 +1,12 @@
+import React, { Component } from 'react';
+
+export default class Home extends Component {
+  render() {
+
+    return (
+      <div>
+        <h1>Home page</h1>
+      </div>
+    );
+  }
+}
