+++ src/pages/Movie.js
@@ -1,0 +1,12 @@
+import React, { Component } from 'react';
+
+export default class Movie extends Component {
+  render() {
+
+    return (
+      <div>
+        <h1>Movie detail page</h1>
+      </div>
+    );
+  }
+}
